@@ -6,6 +6,7 @@
     "node": "12.13.0"
   },
   "dependencies": {
+    "@mikecousins/react-pdf": "^5.5.1",
     "@types/jest": "^24.0.19",
     "@types/node": "^12.11.6",
     "@types/pdfjs-dist": "^2.1.3",
@@ -20,12 +21,12 @@
     "node-sass": "^4.13.0",
     "normalize.css": "^8.0.1",
     "pdfjs-dist": "^2.2.228",
-    "react": "^16.11",
+    "react": "^16.12.0",
     "react-bootstrap": "^1.0.0-beta.14",
     "react-bootstrap-table-next": "^3.3.1",
     "react-bootstrap-table2-editor": "^1.4.0",
     "react-bootstrap-table2-paginator": "^2.1.0",
-    "react-dom": "^16.11",
+    "react-dom": "^16.12.0",
     "react-google-recaptcha": "^2.0.1",
     "react-helmet": "^5.2.1",
     "react-hot-loader": "^4.12.17",
