@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
-import DocumentViewer from './DocumentViewer';
+import DocumentViewer from './DocumentViewerPDFJS';
 import getServerURL from '../serverOverride';
 
 interface Props {
@@ -49,7 +49,13 @@ class ViewDocument extends Component<Props, State> {
     } = this.state;
     return (
       <div>
-        { pdfFile ? <DocumentViewer pdfFile={pdfFile} /> : <div /> }
+        { pdfFile ? 
+          <button type="button">
+          <a target="_blank" href={URL.createObjectURL(pdfFile)}>
+          View Document
+          </a>
+          </button>
+      : <div /> }
         <Link to="/my-documents">
           <button type="button">
             Back
